backend/requirements.txt
# Backend dependencies
# Install with: pip install -r requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
ultralytics>=8.3.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
